requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
defusedxml==0.7.1
Django==3.2.9
django-crispy-forms==1.13.0
django-guardian==2.4.0
django-settings-export==1.2.1
idna==3.3
isodate==0.6.0
Jinja2==2.11.3
lxml==4.6.4
MarkupSafe==2.0.1
mock==4.0.3
onelogin==2.0.2
python-dateutil==2.8.2
python3-saml==1.12.0
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
xmlsec==1.3.12
